requirements.txt
asgiref==3.2.10
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
